requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.25
asyncpg==0.29.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.1
alembic==1.13.1
fastapi[all]>=0.109.2
requests
